Removed .lower() from self._vendor and payload["Vendor"]
Removed `.lower()` from self._vendor and `if payload["Vendor"]  ..`. This shold fix the issue of case mismatching when sending and receiving IR .. This might become breaking change if someone populated its vendor in lower case in configuration.yaml
The vendor now has to be populated exactly as in the 'irreceived IRHvac'. Just like before the bug and just like it is explained in the documentation.