Fix #78 Add new option `dnt_send_list`
This is a setting that does not transmit the data if the On transmission
is a toggle operation of a function, and if tasmota_irhvac sends On each
time it does not result in normal operation. If you want a toggle
action, it can be sent by the service.

dnt_send_list: #optional - default []
    - Beep
    - Clean
    - Econo
    - Filter
    - Light
    - Quiet
    - Sleep
    - SwingH
    - SwingV
    - Turbo