Prettify fan modes for ACs with min/max and without high/low
Home Assistant has standard high/low fan modes that are localized and
displayed in the UI in a pretty way. Tasmota IRHVAC has min, low,
middle, high and max, along with a few other modes. Min and max are
displayed in Home Assistant as is and not localized. While nothing can
be done on the integration side for air conditioners that use 5 levels
of the fan, some air conditioners have 3 levels, but use min/medium/max
instead of low/medium/high. For such ACs, convert min->low and max->high
to display pretty localized values in the UI.