feat: Detecting and setting special modes
Some air conditioners have a function that switches to a special mode,
such as cleaning mode, after operation. Previously, this would be a sign
that the unit was still operating in the previous mode, but now the
system can detect this and allow the user to select the mode to set.